fix: add Exception Handler for DatabaseException and ResourceNotFoundException

+++ src/main/java/com/jciterceros/DSCommerce/services/ProductService.java
@@ -3,10 +3,14 @@ package com.jciterceros.DSCommerce.services;
 import com.jciterceros.DSCommerce.dto.ProductDTO;
 import com.jciterceros.DSCommerce.entities.Product;
 import com.jciterceros.DSCommerce.repositories.ProductRepository;
+import com.jciterceros.DSCommerce.services.exceptions.DatabaseException;
+import com.jciterceros.DSCommerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -51,9 +55,16 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
